AR-1: fix PG build — cast u64 to usize for mark_read/ack match arms
rusqlite::Connection::execute returns usize while
postgres::Client::execute returns u64, causing incompatible match
arm types when compiled with --features pg.

+++ engine/src/storage/local_product_store/agent_runtime.rs
@@ -517,7 +517,7 @@ impl LocalProductStore {
                         &json!({"message_id": message_id}),
                     )?;
                 }
-                Ok(n)
+                Ok(n as usize)
             })?,
         };
         if affected == 0 {
@@ -569,7 +569,7 @@ impl LocalProductStore {
                         &json!({"message_id": message_id}),
                     )?;
                 }
-                Ok(n)
+                Ok(n as usize)
             })?,
         };
         if affected == 0 {
